+++ apps/community/Reports/Controllers/Home.php
@@ -4,6 +4,13 @@ namespace HubletoApp\Community\Reports\Controllers;
 
 class Home extends \HubletoMain\Core\Controller {
 
+  public function getBreadcrumbs(): array
+  {
+    return array_merge(parent::getBreadcrumbs(), [
+      [ 'url' => 'reports', 'content' => $this->translate('Reports') ],
+    ]);
+  }
+
   public function prepareView(): void
   {
     parent::prepareView();