+++ apps/community/Deals/Controllers/Settings.php
@@ -24,9 +24,9 @@ class Settings extends \HubletoMain\Core\Controller
       $this->hubletoApp->setConfigAsBool('showMostValuableDealsInDashboard', $showMostValuableDealsInDashboard);
       $this->hubletoApp->saveConfig('showMostValuableDealsInDashboard', $showMostValuableDealsInDashboard ? '1' : '0');
 
-      $showDealValueByStatusInDashboard = $this->main->urlParamAsBool('showDealValueByStatusInDashboard');
-      $this->hubletoApp->setConfigAsBool('showDealValueByStatusInDashboard', $showDealValueByStatusInDashboard);
-      $this->hubletoApp->saveConfig('showDealValueByStatusInDashboard', $showDealValueByStatusInDashboard ? '1' : '0');
+      $showDealValueByResultInDashboard = $this->main->urlParamAsBool('showDealValueByResultInDashboard');
+      $this->hubletoApp->setConfigAsBool('showDealValueByResultInDashboard', $showDealValueByResultInDashboard);
+      $this->hubletoApp->saveConfig('showDealValueByResultInDashboard', $showDealValueByResultInDashboard ? '1' : '0');
 
       $this->viewParams['settingsSaved'] = true;
     }